Fix remaining verbose compiliation warnings in lib
Most of these are due to:

* unused local variables
* space before method argument list
* space missing between binary operators

In a couple cases where there was an unused local variable that
looks like the expression itself wasn't needed, this removes the
entire line.

In one case, there are duplicate when arguments, such as "x" and ?x.
These were different in Ruby 1.8, and I'm guessing the desired
behavior is either "x" or ?x.ord, since that is the equivalent of
the Ruby 1.8 code.