Default to running the event loop on the main thread
This provides a massive speedup. From the sound of this, you would
think this would cause blocking of the user interface while the
Ruby code runs.  However, that happens regardless of this setting,
so I'm not sure what advantages there are to running the eventloop
on a separate thread.

Fixes #26