Replace Fixnum with Integer
Fixnum was deprecated in Ruby 2.5 and removed in Ruby 3.2. I'm not
sure this is safe, as it is possible Fixnum was used deliberately
instead of Integer, maybe because the integers were going to be
passed to Tcl/Tk.  However, from a partial review, it looks like
the code was just using Fixnum instead of Integer.  For example,
the lib/tkextlib/iwidgets/notebook.rb change passes non-Fixnum
instances directly to tk_send_without_enc, so the only change
is that large negative integers are now changed to nil instead
of being passed directly.